// wow-raid-frontend/src/types/enums.ts
export enum WowClass {
  WARRIOR = 'WARRIOR',
  PALADIN = 'PALADIN',
  HUNTER = 'HUNTER',
  ROGUE = 'ROGUE',
  PRIEST = 'PRIEST',
  DEATH_KNIGHT = 'DEATH_KNIGHT',
  SHAMAN = 'SHAMAN',
  MAGE = 'MAGE',
  WARLOCK = 'WARLOCK',
  MONK = 'MONK',
  DRUID = 'DRUID',
  DEMON_HUNTER = 'DEMON_HUNTER',
  EVOKER = 'EVOKER',
}

export enum WowSpec {
  ARMS = 'ARMS', FURY = 'FURY', PROTECTION_WARRIOR = 'PROTECTION_WARRIOR',
  HOLY_PALADIN = 'HOLY_PALADIN', PROTECTION_PALADIN = 'PROTECTION_PALADIN', RETRIBUTION = 'RETRIBUTION',
  BEAST_MASTERY = 'BEAST_MASTERY', MARKSMANSHIP = 'MARKSMANSHIP', SURVIVAL = 'SURVIVAL',
  ASSASSINATION = 'ASSASSINATION', OUTLAW = 'OUTLAW', SUBTLETY = 'SUBTLETY',
  DISCIPLINE = 'DISCIPLINE', HOLY_PRIEST = 'HOLY_PRIEST', SHADOW = 'SHADOW',
  BLOOD = 'BLOOD', FROST_DK = 'FROST_DK', UNHOLY = 'UNHOLY',
  ELEMENTAL = 'ELEMENTAL', ENHANCEMENT = 'ENHANCEMENT', RESTORATION_SHAMAN = 'RESTORATION_SHAMAN',
  ARCANE = 'ARCANE', FIRE = 'FIRE', FROST_MAGE = 'FROST_MAGE',
  AFFLICTION = 'AFFLICTION', DEMONOLOGY = 'DEMONOLOGY', DESTRUCTION = 'DESTRUCTION',
  BREWMASTER = 'BREWMASTER', MISTWEAVER = 'MISTWEAVER', WINDWALKER = 'WINDWALKER',
  BALANCE = 'BALANCE', FERAL = 'FERAL', GUARDIAN = 'GUARDIAN', RESTORATION_DRUID = 'RESTORATION_DRUID',
  HAVOC = 'HAVOC', VENGEANCE = 'VENGEANCE', DEVOURER = 'DEVOURER',
  DEVASTATION = 'DEVASTATION', PRESERVATION = 'PRESERVATION', AUGMENTATION = 'AUGMENTATION',
}

export enum RaidRole {
  TANK = 'TANK',
  HEALER = 'HEALER',
  DPS = 'DPS',
}

export enum Difficulty {
  NORMAL = 'NORMAL',
  HEROIC = 'HEROIC',
  MYTHIC = 'MYTHIC',
}

export enum RaidStatus {
  OPEN = 'OPEN',
  CLOSED = 'CLOSED',
}

export enum RegistrationStatus {
  CONFIRMED = 'CONFIRMED',
  WAITING = 'WAITING',
  ABSENT = 'ABSENT',
}

export enum UserRole {
  ADMIN = 'ADMIN',
  RAID_LEADER = 'RAID_LEADER',
  MEMBER = 'MEMBER',
}

export enum BoardType {
  NOTICE = 'NOTICE',
  FREE = 'FREE',
}

export enum NotificationType {
  NEW_RAID = 'NEW_RAID',
  REGISTRATION_COMPLETE = 'REGISTRATION_COMPLETE',
  REGISTRATION_CANCELLED = 'REGISTRATION_CANCELLED',
  STATUS_CHANGED = 'STATUS_CHANGED',
  RAID_CLOSED = 'RAID_CLOSED',
  RAID_UPDATED = 'RAID_UPDATED',
}